Move structs to single file
