+++ api/recipes.go
@@ -1,14 +1,5 @@
 package main
 
-type recipe struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-type recipes struct {
-	Recipes []recipe `json:"recipes"`
-}
-
 func getRecipes() (recipes, error) {
 	var recipeList []recipe
 	recipeList = make([]recipe, 2)
